requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
botocore==1.40.18
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
dask==2025.7.0
databricks-sdk==0.65.0
datasets==2.20.0
dill==0.3.8
distributed==2025.7.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.5.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
locket==1.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
open-clip-torch==2.24.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.2
partd==1.4.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray[default]==2.9.3
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
s3fs==2024.5.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.37.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
sympy==1.14.0
tblib==3.1.0
threadpoolctl==3.6.0
timm==1.0.19
tomli==2.2.1
toolz==1.0.0
torch==2.2.0
torchvision==0.17.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.3
wcwidth==0.2.13
webdataset==0.2.57
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
